Revert "WIP: data-tiles are not initialized properly"
This reverts commit 66cc1040df9267fdb7684349fb7a36bf933b4173.

Reasons why I am reverting this:

While adding and extra layer/component to encapsulate all data tiles
logic within Board, the code brearks apart completely due to
**lifetime**:

DataTile define its entities field as a Vector of references to
BoardEntities. The type anotation requires a lifetime to specify
how long the references live.

That is fine if we just work with Rust structs, but GodotClasses
and and thus Rust structs derived from them **do NOT support lifetime
annotation**.

This causes that in the DataTile struct we can use a user-defined lifetime
(`<'a'>`) but as it is propragated and eventually a DataTile is used
from within a GodotClass to add an entity to the `entities` vector
within DataTile, we will have to leverage the `'static'` lifetime.

This ends up making the code imposible to expand as Rust will constantly
complain that a variable/attribute/parameter lifetime will not live
enough as it should match `static`.

As of now I am going to leave the Board with its data field with
lifetime annotation. If it causes the whole code to became unworkable in
the future I will have to get rid of any code/field that requires
lifetimes, most likely I will need to only have structs that derive from
GodotClasses.